add floyd warshall

+++ graphs/graph.py
@@ -10,7 +10,7 @@ class Graph(DiGraph[T]):
         super().__init__(*edges, **kwargs)
 
     @override
-    def add_edge(self, from_v: T, to_v: T, weight: Optional[int] = None)->None:
+    def add_edge(self, from_v: T, to_v: T, weight: Optional[float] = None)->None:
         super().add_edge(from_v, to_v, weight)
         super().add_edge(to_v, from_v, weight)
